+++ src/index.ts
@@ -43,8 +43,9 @@ const {
 
 const server = new McpServer({
     name: "TestRail MCP Server",
-    version: "2.3.0",
+    version: "2.4.0",
 });
+
 
 const client = new TestRailClient(TESTRAIL_INSTANCE_URL, TESTRAIL_USERNAME, TESTRAIL_API_KEY);